+++ sei-ibc-go/modules/core/05-port/types/module.go
@@ -51,6 +51,7 @@ type IBCModule interface {
 		ctx sdk.Context,
 		portID,
 		channelID string,
+		counterpartyChannelID string,
 		counterpartyVersion string,
 	) error