+++ sei-ibc-go/testing/mock/ibc_app.go
@@ -40,6 +40,7 @@ type MockIBCApp struct {
 		ctx sdk.Context,
 		portID,
 		channelID string,
+		counterpartyChannelID string,
 		counterpartyVersion string,
 	) error